Fix some clang-tidy bugprone-narrowing-conversions warnings

+++ src/sound/DACSound8U.cc
@@ -1,6 +1,7 @@
 #include "DACSound8U.hh"
 #include "DeviceConfig.hh"
 #include "MSXMotherBoard.hh"
+#include "narrow.hh"
 
 namespace openmsx {
 
@@ -15,7 +16,7 @@ DACSound8U::DACSound8U(std::string_view name_, static_string_view desc,
 
 void DACSound8U::writeDAC(uint8_t value, EmuTime::param time)
 {
-	DACSound16S::writeDAC(int(value) - 0x80, time);
+	DACSound16S::writeDAC(narrow<int16_t>(narrow<int>(value) - 0x80), time);
 }
 
 } // namespace openmsx
